base: fix race condition with occluded keyguard on non-navbar devices
How to reproduce:
1) select non secure keyguard
2) setup an activity which has FLAG_DISMISS_KEYGUARD set
e.g. DeskClock alarm
3) turn screen off to lock device
4) activity will turn on screen and occlude keyguard
5) turn off screen to lock device
6) turn screen on
7) notice that keyguard layout has a top offset which will only be rest on anogther unlock and lock cycle

This is a rather tricky issue. The problem is that the layout
of StatusBarWindowView will enter a undefined state of keyguard shown
and still occluded when the layout is recaclulated

On devices with navbar this is not an isse cause there a separate
layout step will be done when the navbar is hidden as part of
showKeyguard which will reset the wrong status

On non-navbar devices there is no separate layout step so the
undefined state will be refelected in the layout.
getFitsSystemWindows() is set to true even if the keyguard should always
be shown fullscreen.

There are different ways to solve this e.g. force a "fake" layout
event also for non-navbar devices but the better solution is to prevent
that the undefined state can even be entered. So reset occluded right
when showKeyguard is set to true after occluded was true.

Change-Id: Iac2d4844f065a9405d05a77a51a2fe17d5cbd177
Signed-off-by: AudioGod <audiogod@sonic-developers.com>