bootanimation: Do not free frame map
The frame map is reused when needSaveMem==true, resulting in a double
free and segfault.

The only side effect would seem to be failure to cleanup all resources
at exit.  That's acceptable, as the kernel will do it anyway.

Note that in 12.1, the frame map release commit was reverted by CAF in
commit I535330edb636f9457beed0849147168eaa6ae2ec, which results in the
same behavior.

Change-Id: I4ea66dccd1f5c9d784701242fba8700bce271579
Signed-off-by: AudioGod <audiogod.adi@gmail.com>